Return better output shape for Loop with zero iterations (#1233)
* Attempt to provide the correct rank for an output from a Loop node when there are no iterations.

For a loop output (vs. loop carried dependency) the first dimension is the iteration count so will have a value of 0 and the output size will be zero. Use the rank of the matching subgraph output if available.

If the subgraph output rank is not available output a warning and use a rank 1 shape of {0}.