skip optional inputs for scan subgraphs (#3349)
* skip optional inputs for scan subgraphs

We may have cases where the subgraph has optionial inputs that appear
in both subgraph's input and initializer, but not in the node's input.
In such cases, the input model might be invalid, but let's not choke
on it. Instead, let's issue a warning, skip the optional inputs,
and keep going forward.

* address CR feedback