[OpenVINO-EP] Fixes OpenVINO-EP build on windows (#6726)
* Fixes OpenVINO-EP windows build

Openvino EP build is broken on windows. The issue
is wchar_t is UTF-16 on windows while on other platforms
such as Linux and MacOS, wchar_t is UTF-32.

so wide Unicode string has to be converted to an UTF8 string
for sure on windows.

This commit fixes this issue.