Add  DnnlOpManager (#7521)
* Add  DnnlOpManager

The DnnlOpManager is able to more accurately check if a node is
supported by the DNNLExecutionProvider.

The DNNLExecutionProvider::GetCapability function has been updated
to use the DnnlOpManager.

This commit adds the ability to check if data type, attributes,
and tensor dimensions of the node are supported.

The IsDimensionSupported function is no longer needed since the checks
it was doing have been moved into the individual implementations of
the virtual class DnnlNodeCapability.

Signed-off-by: George Nash <george.nash@intel.com>

* Fix AveragePool entry in the DnnlOpManager

Added check for ceil_mode attribute in the PoolNodeCapability
check.  DnnlExecutionProvider does not support ceil_mode other
than the default value.

Signed-off-by: George Nash <george.nash@intel.com>